% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridl.R
\name{apihelper}
\alias{apihelper}
\alias{ridl}
\title{Helper function to make API calls}
\usage{
ridl(action, ..., .encoding = "json")
}
\arguments{
\item{action}{Operation to execute. See \href{https://docs.ckan.org/en/2.9/api/}{CKAN's API documentation} for details.}

\item{.encoding}{HTTP POST encoding to use - one of \code{json}, \code{form}, or \code{multipart}.}
}
\value{
\code{httr::response} object with the result of the call.
}
\description{
Helper function to make API calls
}
\examples{
#ridl()
}
