% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_list.R
\name{container_list}
\alias{container_list}
\title{container_list}
\usage{
container_list(parent)
}
\arguments{
\item{parent}{name of the container}
}
\value{
catalog of containers
}
\description{
Provide a list of all child containers for a specific container
}
\details{
This function is used to generate a regional dashbaord..
Be carefull - it's an expansive functions at it needs to parse the entire
content of the server....

uses https://docs.ckan.org/en/2.9/api/index.html#ckan.logic.action.get.organization_list
}
\examples{
# catalog <- container_list()
# groups_name <- catalog |>
#                   dplyr::select(groups_name) |>
#                   dplyr::distinct()

}
