% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container_show.R
\name{container_show}
\alias{container_show}
\title{container_show}
\usage{
container_show(id)
}
\arguments{
\item{id}{id or name of the container}
}
\value{
a dataframe with container metadata
}
\description{
Get an overview of accessible infos about all containers in RIDL
Use https://docs.ckan.org/en/2.9/api/index.html#ckan.logic.action.get.organization_show
}
\examples{
# americasdataset <- container_show( id = "americas-regional-dataset")
}
