% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metadata.R
\name{dataset_metadata}
\alias{dataset_metadata}
\title{Convenience function to record dataset metadata}
\usage{
dataset_metadata(
  title = NULL,
  name = NULL,
  short_title = NULL,
  notes = NULL,
  tag_string = NULL,
  url = NULL,
  owner_org = NULL,
  private = NULL,
  visibility = NULL,
  external_access_level = NULL,
  data_sensitivity = NULL,
  original_id = NULL,
  data_collector = NULL,
  date_range_start = NULL,
  date_range_end = NULL,
  keywords = NULL,
  unit_of_measurement = NULL,
  sampling_procedure = NULL,
  operational_purpose_of_data = NULL,
  `hxl-ated` = NULL,
  process_status = NULL,
  identifiability = NULL,
  geog_coverage = NULL,
  data_collection_technique = NULL,
  linked_datasets = NULL,
  archived = NULL,
  admin_notes = NULL,
  sampling_procedure_notes = NULL,
  response_rate_notes = NULL,
  data_collection_notes = NULL,
  weight_notes = NULL,
  clean_ops_notes = NULL,
  data_accs_notes = NULL,
  ddi = NULL,
  ...
)
}
\arguments{
\item{title}{Title(*) - Make sure to include: 'Survey name/title', 'Location', 'Country', and 'Year(s)' in the order indicated.}

\item{name}{URL(*) - The canonical name of the dataset, eg. my-dataset.}

\item{short_title}{Short title - eg. Short title for the project.}

\item{notes}{Description(*) - Some useful notes about the data. Please include the number of observations.}

\item{tag_string}{Tags - eg. economy, mental health, government.}

\item{url}{Project URL - Website URL associated with this data project (if applicable).}

\item{owner_org}{Data container(*) - Use the canonical name for the container.}

\item{private}{Visibility (Private/Public).}

\item{visibility}{Internal Access Level(*). Allowed values: \code{restricted} (Private), \code{public} (Internally Visible).}

\item{external_access_level}{External access level(*). Allowed values: \code{not_available} (Not available), \code{direct_access} (Direct access), \code{public_use} (Public use), \code{licensed_use} (Licensed use), \code{data_enclave} (Data enclave), \code{open_access} (Open access).}

\item{data_sensitivity}{Data sensitivity - Apply to both Anonymized and Personally identifiable data. Allowed values: \code{yes} (Yes), \code{no} (No).}

\item{original_id}{Original ID - If the dataset already has an ID from the source org, DDI, etc...}

\item{data_collector}{Data Collector(*) - Which organization owns / collected the data. Multiple values are allowed.}

\item{date_range_start}{Date collection first date - Use dd/mm/yyyy format.}

\item{date_range_end}{Date collection last date - Use dd/mm/yyyy format.}

\item{keywords}{Topic classifications(*) - Tags useful for searching for the datasets. Multiple values are allowed. See \code{\link{keywords}}}

\item{unit_of_measurement}{Unit of measurement(*) - Unit of measurement / observation for the dataset.}

\item{sampling_procedure}{Sampling Procedure. Multiple values are allowed. Allowed values: \code{total_universe_complete_enumeration} (Total universe/Complete enumeration), \code{probability_simple_random} (Probability: Simple random), \code{probability_systematic_random} (Probability: Systematic random), \code{probability_stratified} (Probability: Stratified), \code{probability_stratified_proportional} (Probability: Stratified: Proportional), \code{probability_stratified_disproportional} (Probability: Stratified: Disproportional), \code{probability_cluster} (Probability: Cluster), \code{probability_cluster_simple_random} (Probability: Cluster: Simple random ), \code{probability_cluster_stratified_random} (Probability: Cluster: Stratified random), \code{probability_multistage} (Probability: Multistage), \code{nonprobability} (Non-probability), \code{nonprobability_availability} (Non-probability: Availability), \code{nonprobability_purposive} (Non-probability: Purposive), \code{nonprobability_quota} (Non-probability: Quota), \code{nonprobability_respondentassisted} (Non-probability: Respondent-assisted), \code{mixed_probability_nonprobability} (Mixed probability and non-probability), \code{other_other} (Use if the sampling procedure is known, but not found in the list..).}

\item{operational_purpose_of_data}{Operational purpose of data - Classification of the type of data contained in the file. Multiple values are allowed. Allowed values: \code{participatory_assessments} (Participatory assessments), \code{baseline_household_survey} (Baseline Household Survey), \code{rapid_needs_assessment} (Rapid Needs Assessment), \code{protection_monitoring} (Protection Monitoring), \code{programme_monitoring} (Programme monitoring), \code{population_data} (Population Data), \code{cartography} (Cartography, Infrastructure & GIS).}

\item{process_status}{Dataset Process Status. Allowed values: \code{raw} (Raw-Uncleaned), \code{cleaned} (Cleaned Only), \code{anonymized} (Cleaned & Anonymized).}

\item{identifiability}{Identifiability. Allowd values: \code{personally_identifiable} (Personally identifiable), \code{anonymized_enclave} (Anonymized 1st level: Data Enclave - only removed direct identifiers), \code{anonymized_scientific} (Anonymized 2st level: Scientific Use File (SUF)), \code{anonymized_public} (Anonymized 3rd level: Public Use File (PUF)).}

\item{geog_coverage}{Geographic Coverage - eg. National coverage, or name of the area, etc.}

\item{data_collection_technique}{Data collection technique(*). Allowed values: \code{nf} (Not specified), \code{f2f} (Face-to-face interview), \code{capi} (Face-to-face interview: Computerised), \code{cami} (Face-to-face interview: Mobile), \code{papi} (Face-to-face interview: Paper-and-pencil), \code{tri} (Telephone interview), \code{eri} (E-mail interview), \code{wri} (Web-based interview: audio-visual technology enabling the interviewer(s) and interviewee(s) to communicate in real time), \code{easi} (Self-administered questionnaire: E-mail), \code{pasi} (Self-administered questionnaire: Paper), \code{sasi} (Self-administered questionnaire: SMS/MMS), \code{casi} (Self-administered questionnaire: Computer-assisted), \code{cawi} (Self-administered questionnaire: Web-based), \code{foc} (Face-to-face focus group), \code{tfoc} (Telephone focus group), \code{obs} (Observation), \code{oth} (Other).}

\item{linked_datasets}{Linked Datasets - Links to other RIDL datasets. It supports multiple selections.}

\item{archived}{Archived(*) - Allows users to indicate if the dataset is archived or active. Allowed values: \code{False} (No), \code{True} (Yes).}

\item{admin_notes}{Admin Notes - General. You can use Markdown formatting here.}

\item{sampling_procedure_notes}{Admin Notes - Sampling Procedure. You can use Markdown formatting here.}

\item{response_rate_notes}{Admin Notes - Response Rate. You can use Markdown formatting here.}

\item{data_collection_notes}{Admin Notes - Data Collection. You can use Markdown formatting here.}

\item{weight_notes}{Admin Notes - Weighting. You can use Markdown formatting here.}

\item{clean_ops_notes}{Admin Notes - Cleaning. You can use Markdown formatting here.}

\item{data_accs_notes}{Admin Notes - Access authority. You can use Markdown formatting here.}

\item{ddi}{DDI.}

\item{...}{ignored.}

\item{`hxl-ated`}{HXL-ated.}
}
\value{
A list with the provided metadata.
}
\description{
Convenience function to record dataset metadata
}
\details{
All arguments are of type character. Fields \code{tag_string}, \code{data_collector}, \code{keywords}, \code{sampling_procedure}, and \code{operational_purpose_of_data} accept vectors of multiple values.

Fields marked with a (*) are required for \code{\link{dataset_create()}} and \code{\link{dataset_update()}} operations.
}
\examples{
#dataset_metadata()
library(riddle)
Sys.setenv(USE_UAT=1)
m <- dataset_metadata(title = "Motor Trend Car Road Tests",
                      name = "mtcars",
                      notes = "The data was extracted from the 1974 Motor Trend 
                      US magazine, and comprises fuel consumption and 10 aspects
                      of automobile design and performance for 32 automobiles 
                      (1973–74 models).",
                      owner_org = "exercise-container",
                      visibility = "public",
                      external_access_level = "open_access",
                      data_collector = "Motor Trend",
                      keywords = keywords[c("Environment", "Other")],
                      unit_of_measurement = "car",
                      data_collection_technique = "oth",
                      archived = "False")
m

}
