% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_tibblify.R
\name{dataset_tibblify}
\alias{dataset_tibblify}
\title{dataset_tibblify}
\usage{
dataset_tibblify(x)
}
\arguments{
\item{x}{dataset as a list}
}
\value{
dataset
}
\description{
Helper function to package API results as a  tibble
}
\examples{
m <- dataset_metadata(title = "Motor Trend Car Road Tests",
                      name = "mtcars",
                      notes = "The data was extracted from the 1974 Motor Trend 
                      US magazine, and comprises fuel consumption and 10 aspects
                      of automobile design and performance for 32 automobiles 
                      (1973–74 models).",
                      owner_org = "exercise-container",
                      visibility = "public",
                      geographies = "UNSPECIFIED",
                      external_access_level = "open_access",
                      data_collector = "Motor Trend",
                      keywords = keywords[c("Environment", "Other")],
                      unit_of_measurement = "car",
                      data_collection_technique = "oth",
                      archived = "False")

m1 <- dataset_tibblify(m)
m1
}
