% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_child_containers.R
\name{find_child_containers}
\alias{find_child_containers}
\title{find_child_containers}
\usage{
find_child_containers(parent, catalog)
}
\arguments{
\item{parent}{name of the parent container}

\item{catalog}{daaframe object with a catalog of container produced by container_list()}
}
\value{
vector with all child container
}
\description{
Provide a list of all child containers - including nested one - for a specific container
}
\details{
Be carefull - it's an expansive functions at it needs to parse the entire
content of the server....
}
\examples{
#catalog <- container_list()
# containerAmericas <- find_child_containers(parent = "americas",
#                                            catalog = catalog)
}
