% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-keywords.R
\docType{data}
\name{keywords}
\alias{keywords}
\title{Package keywords}
\format{
A named character vector mapping user-visible labels (the names) to their corresponding codes in the system (the values).
}
\usage{
keywords
}
\description{
As extracted from the \href{https://github.com/okfn/ckanext-unhcr/blob/master/ckanext/unhcr/schemas/dataset.json}{dataset schema}.
}
\keyword{datasets}
