% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_fetch.R
\name{resource_fetch}
\alias{resource_fetch}
\title{Fetch resource from RIDL}
\usage{
resource_fetch(url, path = tempfile(), uat = FALSE)
}
\arguments{
\item{url}{The URL of the resource to fetch}

\item{path}{Location to store the resource}

\item{uat}{Boolean TRUE /FALSE tells whether to use
https://ridl-uat.unhcr.org/ or https://ridl.unhcr.org/.
FALSE per default}
}
\value{
Path to the downloaded file
}
\description{
Fetch resource from RIDL
}
\examples{
# resource_fetch()
}
