% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resource_metadata}
\alias{resource_metadata}
\title{Convenience function to record resource metadata}
\usage{
resource_metadata(
  type = NULL,
  url = NULL,
  name = NULL,
  description = NULL,
  format = NULL,
  file_type = NULL,
  date_range_start = NULL,
  date_range_end = NULL,
  version = NULL,
  `hxl-ated` = NULL,
  process_status = NULL,
  identifiability = NULL,
  ...
)
}
\arguments{
\item{type}{Resource type(*) - The kind of file you want to upload. Allowed values: \code{data} (Data file), \code{attachment} (Additional attachment).}

\item{url}{Upload - The file name as it will be recorded in the system.}

\item{description}{Description - Some usefule notes about the data.}

\item{format}{File format - eg. CSV, XML, or JSON.}

\item{file_type}{File type(*) - Indicates what is contained in the file. Allowed values: \code{microdata} (Microdata), \code{questionnaire} (Questionnaire), \code{report} (Report), \code{sampling_methodology} (Sampling strategy & methodology Description), \code{infographics} (Infographics & Dashboard), \code{script} (Script), \verb{concept note} (Concept Note), \code{other} (Other).}

\item{date_range_start}{Data collection first date(*) - Use yyyy-mm-dd format.}

\item{date_range_end}{Data collection last date(*) - Use yyyy-mm-dd format.}

\item{version}{Version(*).}

\item{process_status}{File process status(*) - Indicates the processing stage of the data. 'Raw' means that the data has not been cleaned since collection. 'In process' means that it is being cleaned. 'Final' means that the dataset is final and ready for use in analytical products. Allowed valued: \code{raw} (Raw-Uncleaned), \code{cleaned} (Cleaned Only), \code{anonymized} (Cleaned & Anonymized).}

\item{identifiability}{Identifiability(*) - Indicates if personally identifiable data is contained in the dataset. Allowed values: \code{personally_identifiable} (Personally identifiable), \code{anonymized_enclave} (Anonymized 1st level: Data Enclave - only removed direct identifiers), \code{anonymized_scientific} (Anonymized 2st level: Scientific Use File (SUF)), \code{anonymized_public} (Anonymized 3st level: Public Use File (PUF)).}

\item{...}{ignored.}

\item{multipart_name}{Name - eg. January 2011 Gold Prices.}

\item{`hxl-ated`}{HXL-ated. Allowed values: \code{False} (No), \code{True} (Yes).}
}
\value{
A list with the provided metadata.
}
\description{

}
\details{
All arguments are of type character.

Fields marked with a (*) are required for \code{\link{resource_create()}} and \code{\link{resource_update()}} operations.
}
