% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_tibblify.R
\name{resource_tibblify}
\alias{resource_tibblify}
\title{resource_tibblify}
\usage{
resource_tibblify(x)
}
\arguments{
\item{x}{list}
}
\value{
list tiblified
}
\description{
Helper function to package API results as a  tibble
}
\examples{
m <- riddle::resource_metadata(type = "data",
                       url = "mtcars.csv",
                       name = "mtcars.csv",
                       format = "csv",
                       file_type = "microdata",
                       date_range_start = "1973-01-01",
                       date_range_end = "1973-12-31",
                       version = "1",
                       visibility = "public",
                       process_status = "raw",
                       identifiability = "anonymized_public")
m1 <- riddle::resource_tibblify(m)
m1
}
