% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riddle-package.R
\docType{package}
\name{riddle-package}
\alias{riddle}
\alias{riddle-package}
\title{riddle: Minimal Client for the UNHCR RIDL Platform}
\description{
Wrapper functions to work with the RIDL API from R
}
\author{
\strong{Maintainer}: Hisham Galal \email{galal@unhcr.org}

}
\keyword{internal}
