% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\alias{dataset_create}
\alias{dataset_show}
\alias{dataset_update}
\alias{dataset_patch}
\alias{dataset_delete}
\title{Work with RIDL datasets (datasets)}
\usage{
dataset_create(metadata)

dataset_show(id)

dataset_update(id, metadata)

dataset_patch(id, metadata)

dataset_delete(id)
}
\arguments{
\item{metadata}{Metadata created by \code{\link{dataset_metadata()}}.}

\item{id}{The id or name of the dataset.}
}
\value{
The dataset.
}
\description{
Work with RIDL datasets (datasets)
}
\details{
You must have the necessary permissions to create, edit, or delete datasets.

Note that several fields are required for \code{dataset_create()} and \code{dataset_update()} operations to succeed. Consult \code{\link{dataset_metadata()}} for the details.

For \code{dataset_update()}/\code{dataset_patch()} operations, it is recommended to call \code{dataset_show()}, make the desired changes to the result, and then call \code{dataset_update()}/\code{dataset_patch()} with it.

The difference between the update and patch methods is that the patch will perform an update of the provided parameters, while leaving all other parameters unchanged, whereas the update methods deletes all parameters not explicitly provided in the \code{metadata}.
}
\examples{

library(riddle)
Sys.setenv(USE_UAT=1)
#Sys.unsetenv("USE_UAT")

### TO FIX -- geographies: Missing value
## __type: Validation Error ---
## cf https://github.com/okfn/ckanext-unhcr/blob/master/ckanext/unhcr/schemas/dataset.json#L670:L682

# m <- dataset_metadata(title = "Motor Trend Car Road Test two",
#                       name = "mtcars",
#                       notes = "The data was extracted from the 1974 Motor Trend
#                       US magazine, and comprises fuel consumption and 10 aspects
#                       of automobile design and performance for 32 automobiles
#                       (1973–74 models).",
#                       owner_org = "exercise-container",
#                       visibility = "public",
#                       geographies = "UNSPECIFIED",
#                       external_access_level = "open_access",
#                       data_collector = "Motor Trend",
#                       keywords = keywords[c("Environment", "Other")],
#                       unit_of_measurement = "car",
#                       data_collection_technique = "oth",
#                       archived = "False")
# ## For the above to work - you need to make sure you have at least editor access
# # to the corresponding container - i.e. owner_org = "exercise-container"
# p <- dataset_create(m)
# The return value is a representation of the dataset we just created in
# RIDL that you could inspect like any other R object.
#p



#dataset_show("unhcr-cbi-americas-quarterly-report")

}
