% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\alias{dataset_create}
\alias{dataset_show}
\alias{dataset_update}
\alias{dataset_patch}
\alias{dataset_delete}
\title{Work with RIDL datasets (datasets)}
\usage{
dataset_create(metadata)

dataset_show(id)

dataset_update(id, metadata)

dataset_patch(id, metadata)

dataset_delete(id)
}
\arguments{
\item{metadata}{Metadata created by \code{\link{dataset_metadata()}}.}

\item{id}{The id or name of the dataset.}
}
\value{
The dataset.
}
\description{
Work with RIDL datasets (datasets)
}
\details{
You must have the necessary permissions to create, edit, or delete datasets.

Note that several fields are required for \code{dataset_create()} and \code{dataset_update()} operations to succeed.
Consult \code{\link{dataset_metadata()}} for the details.

For \code{dataset_update()}/\code{dataset_patch()} operations, it is recommended to
call \code{dataset_show()}, make the desired changes to the result,
and then call \code{dataset_update()}/\code{dataset_patch()} with it.

The difference between the update and patch methods is that the patch will
perform an update of the provided parameters, while leaving all other
parameters unchanged, whereas the update methods deletes all parameters
not explicitly provided in the \code{metadata}.
}
\examples{


#-----
# test search in prod
Sys.unsetenv("USE_UAT")
#Sys.setenv(USE_UAT=0)
# riddle::dataset_show(id = "unhcr-cbi-americas-quarterly-report")
# 
# p <- riddle::dataset_show('rms_v4')
# list_of_ressources <- p[["resources"]][[1]]
# list_of_ressources



#-----
# Test create in UAT
Sys.setenv(USE_UAT=1)

# p <- riddle::dataset_show('testedouard2')

### TO FIX -- geographies: Missing value
## __type: Validation Error ---
## cf https://github.com/okfn/ckanext-unhcr/blob/master/ckanext/unhcr/schemas/dataset.json#L670:L682

m <- riddle::dataset_metadata(title = "Motor Trend Car Road Test two",
                      name = "mtcars_ed",
                      notes = "The data was extracted from the 1974 Motor Trend
                      US magazine, and comprises fuel consumption and 10 aspects
                      of automobile design and performance for 32 automobiles
                      (1973–74 models).",
                      owner_org = "Americas",
                      visibility = "public",
                      geographies = "UNSPECIFIED",
                      external_access_level = "open_access",
                      data_collector = "Motor Trend",
                      keywords = keywords[c("Environment", "Other")],
                      unit_of_measurement = "car",
                      data_collection_technique = "oth",
                      archived = "False")
# ## For the above to work - you need to make sure you have at least editor access
# to the corresponding container - i.e. owner_org = "exercise-container"
#p <- riddle::dataset_create(metadata = m)
# The return value is a representation of the dataset we just created in
# RIDL that you could inspect like any other R object.
#p

#-----
# Test create in prod
Sys.unsetenv("USE_UAT")
# m1 <- riddle::dataset_metadata(title = "Test",
#                       name = "Test",
#                       notes = "The data was extracted from kobo.",
#                       owner_org = "Americas : Regional Datasets",
#                       visibility = "public",
#                       geographies = "UNSPECIFIED",
#                       external_access_level = "open_access",
#                       data_collector = "UNHCR",
#                       keywords = keywords[c("Environment", "Other")],
#                       unit_of_measurement = "car",
#                       data_collection_technique = "oth",
#                       archived = "False")
# p <- riddle::dataset_create(metadata = m1)


}
