% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/search.R
\name{resource}
\alias{resource}
\alias{resource_create}
\alias{resource_update}
\alias{resource_patch}
\alias{resource_delete}
\title{Work with RIDL resources (files)}
\usage{
resource_create(pkgid, res_metadata)

resource_update(id, res_metadata)

resource_patch(id, res_metadata)

resource_delete(id)

resource_create(pkgid, res_metadata)

resource_update(id, res_metadata)

resource_patch(id, res_metadata)

resource_delete(id)
}
\arguments{
\item{pkgid}{The id or name of the dataset to which this resource belongs to.}

\item{res_metadata}{Metadata created by \code{\link{resource_metadata()}}.}

\item{id}{The id or name of the resource.}
}
\value{
The resource.
}
\description{
Work with RIDL resources (files)
}
\details{
You must have the necessary permissions to create, edit, or delete
datasets and their resources.

Note that several fields are required for \code{resource_create()} and
\code{resource_update()} operations to succeed.
Consult \code{\link{resource_metadata()}} for the details.

For \code{resource_update()}/\code{resource_patch()} operations, it is recommended to
call \code{resource_show()}, make the desired changes to the result, and then
call \code{resource_update()}/\code{resource_patch()} with it.

The difference between the update and patch methods is that the patch will
perform an update of the provided parameters, while leaving all other
parameters unchanged, whereas the update methods deletes all parameters
not explicitly provided in the \code{metadata}.
}
\examples{
## let's get again the details of the dataset we want to add the resource in..
#-----
# Test search in prod
Sys.unsetenv("USE_UAT")
p <- riddle::dataset_search("rms_v4")
p

#-----
# Test search in uat
Sys.setenv(USE_UAT=1)
p <- riddle::dataset_search("tests")
p

ridlid <- as.character(p[1, c("id")])

m <- riddle::resource_metadata(type = "data",
                       url = "mtcars.csv",
                       name = "mtcarsriddle",
                       format = "csv",
                       file_type = "microdata",
                       date_range_start = "1973-01-01",
                       date_range_end = "1973-12-31",
                       version = "1",
                       visibility = "public",
                       process_status = "raw",
                       identifiability = "anonymized_public",
upload = httr::upload_file(system.file("extdata/mtcars.csv", package = "readr")))

## let's get again the details of the dataset we want to add the resource in..
#r <- riddle::resource_create(pkgid = ridlid, res_metadata = m)
# Like before, the return value is a tibble representation of the resource.
#r
 
## and now can search for it - checking it is correctly there... 
##riddle::resource_search("name:mtcarsriddle")

# And once we’re done experimenting with the API, we should take down our
# toy dataset since we don’t really need it on RIDL.
#riddle::dataset_delete(p$id)

}
