% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resource}
\alias{resource}
\alias{resource_create}
\alias{resource_upload}
\alias{resource_update}
\alias{resource_patch}
\alias{resource_delete}
\title{Work with RIDL resources (files)}
\usage{
resource_create(pkgid, metadata)

resource_upload(id, path)

resource_update(id, metadata)

resource_patch(id, metadata)

resource_delete(id)
}
\arguments{
\item{pkgid}{The id or name of the package to which this resource belongs to.}

\item{metadata}{Metadata created by \code{\link{resource_metadata()}}.}

\item{id}{The id or name of the resource.}
}
\value{
The resource.
}
\description{
Work with RIDL resources (files)
}
\details{
You must have the necessary permissions to create, edit, or delete packages and their resources.

Note that several fields are required for \code{resource_create()} and \code{resource_update()} operations to succeed. Consult \code{\link{resource_metadata()}} for the details.

For \code{resource_update()}/\code{resource_patch()} operations, it is recommended to call \code{resource_show()}, make the desired changes to the result, and then call \code{resource_update()}/\code{resource_patch()} with it.

The difference between the update and patch methods is that the patch will perform an update of the provided parameters, while leaving all other parameters unchanged, whereas the update methods deletes all parameters not explicitly provided in the \code{metadata}.
}
