% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_fetch.R
\name{resource_fetch}
\alias{resource_fetch}
\title{Fetch resource from RIDL}
\usage{
resource_fetch(url, path = tempfile())
}
\arguments{
\item{url}{The URL of the resource to fetch}

\item{path}{Location to store the resource}
}
\value{
Path to the downloaded file
}
\description{
Fetch resource from RIDL
}
\examples{
# resource_fetch()
}
