% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_fetch.R
\name{resource_fetch}
\alias{resource_fetch}
\title{Fetch resource from RIDL}
\usage{
resource_fetch(url, path = tempfile())
}
\arguments{
\item{url}{The URL of the resource to fetch}

\item{path}{Location to store the resource}
}
\value{
Path to the downloaded file
}
\description{
Fetch resource from RIDL
}
\examples{
#-----
# Test search in prod
#Sys.unsetenv("USE_UAT")


## Example 1: with a direct URL
# resource_fetch(url = 'https://ridl.unhcr.org/dataset/a60f4b79-8acc-4893-8fb9-d52f94416b19/resource/daa2b9e4-bf97-4302-86a5-08bb62a5a937/download/df_age_2022.csv',
#                        path = tempfile())

## Example 2: Let's try to identify a resource - then fetch it locally and update it back... as from here
# https://github.com/unhcr-americas/darien_gap_human_mobility/blob/main/report.Rmd#L38

# p <- riddle::dataset_show('rms_v4')
# ## let's get the fifth resource
# test_ressources <- p[["resources"]][[1]] |> dplyr::slice(5)
#
# ## Get the resource locally..
#resource_fetch(url = test_ressources$url,   path =  here::here("file"))
#
# # Rebuild the metadata
# m <- resource_metadata(type = test_ressources$type, #"data",
#                          url = "df_gender_2020.csv",
# upload = httr::upload_file(here::here("data.csv")),
# ## # Error:
# # ! All columns in a tibble must be vectors.
# # ✖ Column `upload` is a `form_file` object.
#                          name = test_ressources$name, # "Irregular entries by gender in 2022",
#                          format = test_ressources$format, #"csv",
#                          file_type =  test_ressources$file_type, #"microdata",
#                          visibility = test_ressources$visibility, # "public",
#                          date_range_start =  test_ressources$date_range_start, # "2022-01-01",
#                          date_range_end = test_ressources$date_range_end, #as.character(floor_date(today('America/Panama'), "month") - days(1)), #end day of last month
#                          version = test_ressources$version, # "0",
#                          process_status = test_ressources$process_status, #"anonymized",
#                          identifiability = test_ressources$identifiability, #"anonymized_public"
#   )

## Update back -- Does not work...
#r <- resource_update(id = test_ressources$id,  res_metadata = m)
### Error when using the upload...
# No encoding supplied: defaulting to UTF-8.
# Error in !r$success : invalid argument type

}
