% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridl.R
\name{ridl}
\alias{ridl}
\title{apihelper}
\usage{
ridl(action, ..., .encoding = "json", verbose = FALSE)
}
\arguments{
\item{action}{Operation to execute. See \href{https://docs.ckan.org/en/2.9/api/}{CKAN's API documentation} for details.}

\item{...}{whatever is needed}

\item{.encoding}{HTTP POST encoding to use - one of \code{json}, \code{form}, or \code{multipart}.}

\item{verbose}{TRUE FALSE to display info on the console about the API call}
}
\value{
\code{httr::response} object with the result of the call.
}
\description{
Helper function to make \href{https://docs.ckan.org/en/2.9/api/index.html#action-api-reference}{API calls}.
Calls includes the 10 following actions:
}
\details{
On dataset
\itemize{
\item "package_create"
\item "package_update"
\item "package_patch"
\item "package_delete"
\item "package_search"
}

On resource
\itemize{
\item "resource_create"
\item "resource_update"
\item "resource_patch"
\item "resource_delete"
\item "resource_search"
}

The package works with both the production and UAT instances of RIDL.
To use the UAT version, run \code{Sys.setenv(USE_UAT=1)} before calling any functions
from the package.
To go back to the production instance, call \code{Sys.unsetenv("USE_UAT")}.
}
\examples{
# ridl(action ="package_search", as.list("cbi"))
}
