% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/resource.R
\name{search}
\alias{search}
\alias{package_search}
\alias{resource_search}
\title{Searches for packages and resources satisfying a given criteria.}
\usage{
package_search(q = NULL, rows = NULL, start = NULL)

resource_search(query = NULL, rows = NULL, start = NULL)
}
\arguments{
\item{q, query}{The search query.}

\item{rows}{The maximum number of matching rows (datasets) to return. (optional, default: 10, upper limit: 1000)}

\item{start}{The offset in the complete result for where the set of returned datasets should begin.}
}
\value{
A tibble with the search results.
}
\description{
Searches for packages and resources satisfying a given criteria.
}
