% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_search.R, R/resource.R
\name{search}
\alias{search}
\alias{dataset_search}
\alias{resource_search}
\title{Searches for datasets and resources satisfying a given criteria.}
\usage{
dataset_search(q = NULL, rows = NULL, start = NULL)

resource_search(query = NULL, rows = NULL, start = NULL)
}
\arguments{
\item{q, query}{The search query.}

\item{rows}{The maximum number of matching rows (datasets) to return.
(optional, default: 10, upper limit: 1000)}

\item{start}{The offset in the complete result for where the set of returned
datasets should begin.}
}
\value{
A tibble with the search results.

tibble with list of related resource.
}
\description{
Searches for datasets and resources satisfying a given criteria.
}
\examples{
#-----
# Test search in prod
Sys.unsetenv("USE_UAT")
# p <- dataset_search(q = "cbi")
# p



#-----
# Test create in UAT
Sys.setenv(USE_UAT=1)
# p2 <- dataset_search(q = "testedouard2")

}
