% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_report.R
\name{summary_report}
\alias{summary_report}
\title{Generate a RIDL factsheet}
\usage{
summary_report(container = "Americas")
}
\arguments{
\item{container}{list of container to generate the factsheet to generate}
}
\description{
Generate a RIDL factsheet
}
\examples{
# summary_report(year = 2022,
#                                    region = "Americas")
}
